\name{addition}
\alias{addition}
\title{Addition}
\description{
  Zwei Zahlen addieren und das Ergebnis ausgeben.
}
\usage{
  addition(a,b)
}
\arguments{
  \item{a}{
  erste Zahl
}
\item{b}{
  zweite Zahl
}
}
\examples{
  addition(1, 2)     # Should return 3
  addition(10, 5)    # Should return 15
  addition(0, 10)    # Should return 0
}
