\name{square}
\alias{square}
\title{Square a number}
\description{
  Takes in any numeric value and squares it
}
\usage{
  square(x)
}

\arguments{
  \item{x}{
    a numeric value to be squared
  }
}
\examples{
square(2)       # Should return 4
square(4)       # Should return 16
}
